+++ graphix/optimization.py
@@ -419,19 +419,14 @@ class StandardizedPattern(_StandardizedPattern):
             - There cannot be any empty layers.
         """
         oset = frozenset(self.output_nodes)  # First layer by convention.
-        pre_measured_nodes = set(self.results.keys())  # Not included in the partial order layers.
+        pre_measured_nodes = self.results.keys()  # Not included in the partial order layers.
         excluded_nodes = oset | pre_measured_nodes
 
-        zero_indegree = set(self.input_nodes) - excluded_nodes
+        zero_indegree = set(self.input_nodes).union(n.node for n in self.n_list) - excluded_nodes
         dag: dict[int, set[int]] = {
             node: set() for node in zero_indegree
         }  # `i: {j}` represents `i -> j` which means that node `i` must be measured before node `j`.
         indegree_map: dict[int, int] = {}
-
-        for n in self.n_list:
-            if n.node not in oset:  # Pre-measured nodes only appear in domains.
-                dag[n.node] = set()
-                zero_indegree.add(n.node)
 
         def process_domain(node: Node, domain: AbstractSet[Node]) -> None:
             for dep_node in domain:
@@ -487,15 +482,15 @@ class StandardizedPattern(_StandardizedPattern):
         In general, there may exist various layerings which represent the corrections of the pattern. To ensure that a given layering is compatible with the pattern's induced correction function, the partial order must be extracted from a standardized pattern. Commutation of entanglement commands with X and Z corrections in the standardization procedure may generate new corrections, which guarantees that all the topological information of the underlying graph is encoded in the extracted partial order.
         """
         correction_function: dict[int, set[int]] = defaultdict(set)
-        pre_measured_nodes = set(self.results.keys())  # Not included in the flow.
+        pre_measured_nodes = self.results.keys()  # Not included in the flow.
 
         for m in self.m_list:
             if m.plane in {Plane.XZ, Plane.YZ}:
                 raise FlowGenericError(FlowGenericErrorReason.XYPlane)
-            correction_function = _update_corrections(m.node, m.s_domain - pre_measured_nodes, correction_function)
+            _update_corrections(m.node, m.s_domain - pre_measured_nodes, correction_function)
 
         for node, domain in self.x_dict.items():
-            correction_function = _update_corrections(node, domain - pre_measured_nodes, correction_function)
+            _update_corrections(node, domain - pre_measured_nodes, correction_function)
 
         og = (
             self.extract_opengraph()
@@ -527,19 +522,18 @@ class StandardizedPattern(_StandardizedPattern):
 
         Notes
         -----
-        This method makes use of :func:`StandardizedPattern.extract_partial_order_layers` which computes the pattern's direct acyclical graph (DAG) induced by the corrections and returns a particular layer stratification (obtained by doing a topological sort on the DAG). Further, it constructs the pattern's induced correction function from :math:`M` and :math:`X` commands.
-        In general, there may exist various layerings which represent the corrections of the pattern. To ensure that a given layering is compatible with the pattern's induced correction function, the partial order must be extracted from a standardized pattern. Commutation of entanglement commands with X and Z corrections in the standardization procedure may generate new corrections, which guarantees that all the topological information of the underlying graph is encoded in the extracted partial order.
+        The notes provided in :meth:`extract_causal_flow` apply here as well.
         """
-        correction_function: dict[int, set[int]] = defaultdict(set)
-        pre_measured_nodes = set(self.results.keys())  # Not included in the flow.
+        correction_function: dict[int, set[int]] = {}
+        pre_measured_nodes = self.results.keys()  # Not included in the flow.
 
         for m in self.m_list:
             if m.plane in {Plane.XZ, Plane.YZ}:
-                correction_function[m.node].add(m.node)
-            correction_function = _update_corrections(m.node, m.s_domain - pre_measured_nodes, correction_function)
+                correction_function.setdefault(m.node, set()).add(m.node)
+            _update_corrections(m.node, m.s_domain - pre_measured_nodes, correction_function)
 
         for node, domain in self.x_dict.items():
-            correction_function = _update_corrections(node, domain - pre_measured_nodes, correction_function)
+            _update_corrections(node, domain - pre_measured_nodes, correction_function)
 
         og = (
             self.extract_opengraph()
@@ -547,7 +541,7 @@ class StandardizedPattern(_StandardizedPattern):
         partial_order_layers = (
             self.extract_partial_order_layers()
         )  # Raises a `ValueError` if the pattern corrections form closed loops.
-        gf = GFlow(og, dict(correction_function), partial_order_layers)
+        gf = GFlow(og, correction_function, partial_order_layers)
         gf.check_well_formed()
         return gf
 
@@ -645,9 +639,7 @@ def _incorporate_pauli_results_in_domain(
     return odd_outcome == 1, new_domain
 
 
-def _update_corrections(
-    node: Node, domain: AbstractSet[Node], correction: dict[Node, set[Node]]
-) -> dict[Node, set[Node]]:
+def _update_corrections(node: Node, domain: AbstractSet[Node], correction: dict[Node, set[Node]]) -> None:
     """Update the correction mapping by adding a node to all entries in a domain.
 
     Parameters
@@ -660,19 +652,12 @@ def _update_corrections(
         A mapping from measured nodes to sets of nodes on which corrections are applied. This
         dictionary is modified in place.
 
-    Returns
-    -------
-    dict[Node, set[Node]]
-        The updated correction dictionary with `node` added to the correction
-        sets of all nodes in `domain`.
-
     Notes
     -----
     This function is used to extract the correction function from :math:`X`, :math:`Z` and :math:`M` commands when constructing a flow.
     """
     for measured_node in domain:
-        correction[measured_node].add(node)
-    return correction
+        correction.setdefault(measured_node, set()).add(node)
 
 
 def incorporate_pauli_results(pattern: Pattern) -> Pattern: